NXDRIVE-1125: Some cleanup
