+++ nuxeo-drive-client/nxdrive/utils.py
@@ -7,6 +7,7 @@ import re
 import sys
 import time
 import unicodedata
+import urllib
 import urlparse
 from distutils.version import StrictVersion
 from logging import getLogger
@@ -562,6 +563,61 @@ def simplify_url(url):
         new_parts[1] = parts.netloc[:-4]
 
     return urlparse.urlunsplit(new_parts).rstrip('/')
+
+
+def parse_protocol_url(url_string):
+    """Parse URL for which nxdrive is registered as a protocol handler
+
+    Return None if url_string is not a supported URL pattern or raise a
+    ValueError is the URL structure is invalid.
+    """
+    if not url_string.startswith('nxdrive://'):
+        return None
+
+    protocol_regex = [('nxdrive://(?P<cmd>edit)/(?P<scheme>\w*)/'
+                       '(?P<server>.*)/user/(?P<username>\w*)/repo/'
+                       '(?P<repo>\w*)/nxdocid/(?P<docid>(\d|[a-f]|-)*)/'
+                       'filename/(?P<filename>[^/]*)(/downloadUrl/'
+                       '(?P<download>.*)|)'),
+
+                      'nxdrive://(?P<cmd>share_link)/(?P<path>.*)',
+
+                      'nxdrive://(?P<cmd>access)/(?P<path>.*)']
+
+    parsed_url = None
+    for regex in protocol_regex:
+        parsed_url = re.match(regex, url_string)
+        if parsed_url:
+            break
+
+    if not parsed_url:
+        raise ValueError(
+            'Unsupported command {!r} in protocol handler'.format(url_string))
+
+    parsed_url = parsed_url.groupdict()
+    cmd = parsed_url.get('cmd')
+    if cmd == 'edit':
+        return parse_edit_protocol(parsed_url, url_string)
+    if cmd in ('access', 'share_link'):
+        return dict(command=cmd, filepath=parsed_url.get('path'))
+
+
+def parse_edit_protocol(parsed_url, url_string):
+    """ Parse a nxdrive://edit URL for quick editing of nuxeo documents. """
+    scheme = parsed_url.get('scheme')
+    if scheme not in ('http', 'https'):
+        raise ValueError(
+            'Invalid command {} : scheme should be http or https'.format(
+                url_string))
+
+    server_url = '{}://{}'.format(scheme, parsed_url.get('server'))
+
+    return dict(command='download_edit', server_url=server_url,
+                user=parsed_url.get('username'),
+                repo=parsed_url.get('repo'),
+                doc_id=parsed_url.get('docid'),
+                filename=parsed_url.get('filename'),
+                download_url=parsed_url.get('download'))
 
 
 class ServerLoader(object):
