+++ tools/osx/drive/drive/AppDelegate.swift
@@ -2,7 +2,7 @@
 //  AppDelegate.swift
 //  drive
 //
-//  Created by Léa Klein on 27/02/2018.
+//  Created by Léa Klein on 28/02/2018.
 //  Copyright © 2018 Nuxeo. All rights reserved.
 //
